Move photo endpoint to students controller
+ this reduces the amount of information a developer needs to send a valid request for a student photo -- the developer doesn't need to know the photo ID, just the student ID
+ the endpoint sends down the most recent photo
